backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
python-dotenv==1.0.0
pydantic==2.3.0
pytesseract==0.3.10
pdf2image==1.16.3
numpy==1.26.2
Pillow==10.1.0
pandas>=2.0.0
matplotlib>=3.7.0
tk 